+++ test/test_common_extensions.py
@@ -21,7 +21,7 @@ class TestMagic(unittest.TestCase):
 \x56\x01\x29\x00\x46\x4D\x53\x4E\x56\x6D\x70\x34\x32"
         self.expect_ext = ".mp4"
         self.expect_mime = "video/mp4"
-    
+
     def test_file(self):
         """File identification                          |"""
         mp4file = NamedTemporaryFile(delete=False)
@@ -30,16 +30,28 @@ class TestMagic(unittest.TestCase):
         ext = puremagic.from_file(mp4file.name)
         os.unlink(mp4file.name)
         self.assertEqual(self.expect_ext, ext)
-        
+
     def test_hex_string(self):
         """Hex string identification                    |"""
         ext = puremagic.from_string(self.mp4magic)
         self.assertEqual(self.expect_ext, ext)
-        
+
     def test_string(self):
         """String identification                        |"""
         ext = puremagic.from_string(bytes(self.mp4magic))
         self.assertEqual(self.expect_ext, ext)
+
+    def test_string_with_filename_hint(self):
+        """String identification with filename hint     |"""
+        filename=os.path.join(OFFICE_DIR, "test.xlsx")
+        with open(filename, "rb") as f:
+            data = f.read()
+        ext = puremagic.from_string(data)
+        # .docx and .xlsx have same signature
+        self.assertEqual(".docx", ext)
+        # with the hint from_string() shoud find the correct extension
+        ext = puremagic.from_string(data, filename=filename)
+        self.assertEqual(".xlsx", ext)
 
     def test_string_with_confidence(self):
         """String identification: magic_string          |"""
@@ -47,13 +59,22 @@ class TestMagic(unittest.TestCase):
         self.assertEqual(self.expect_ext, ext[0][0])
         self.assertRaises(ValueError, puremagic.magic_string, "")
 
+    def test_magic_string_with_filename_hint(self):
+        """String identification: magic_string with hint|"""
+        filename=os.path.join(OFFICE_DIR, "test.xlsx")
+        with open(filename, "rb") as f:
+            data = f.read()
+        ext = puremagic.magic_string(data, filename=filename)
+        self.assertEqual(".xlsx", ext[0][0])
+
     def test_not_found(self):
         """Bad file type via string                     |"""
         try:
             with self.assertRaises(puremagic.PureError):
                 puremagic.from_string("not applicable string")
         except TypeError:
-            # Python 2.6 doesn't support using assertRaises as a context manager
+            # Python 2.6 doesn't support using
+            # assertRaises as a context manager
             pass
 
     def test_magic_file(self):
@@ -61,7 +82,8 @@ class TestMagic(unittest.TestCase):
         self.assertEqual(puremagic.magic_file(TGA_FILE)[0][0], ".tga")
         open("test_empty_file", "w").close()
         try:
-            self.assertRaises(ValueError, puremagic.magic_file, "test_empty_file")
+            self.assertRaises(ValueError,
+                              puremagic.magic_file, "test_empty_file")
         finally:
             os.unlink("test_empty_file")
 
@@ -92,7 +114,7 @@ class TestMagic(unittest.TestCase):
             self.assertTrue(item.endswith(ext),
                             "Expected .{0}, got {1}".format(item.split(".")[1],
                                                             ext))
-                
+
     def test_audio(self):
         """Test common audio formats                    |"""
         for item in os.listdir(AUDIO_DIR):
@@ -104,10 +126,10 @@ class TestMagic(unittest.TestCase):
             self.assertTrue(item.endswith(ext),
                             "Expected .{0}, got {1}".format(item.split(".")[1],
                                                             ext))
-                
+
     def test_office(self):
         """Test common office document formats          |"""
-        ### Office files have very similar magic numbers, and may overlap
+        # Office files have very similar magic numbers, and may overlap
         for item in os.listdir(OFFICE_DIR):
             puremagic.from_file(os.path.join(OFFICE_DIR, item))
 
@@ -122,7 +144,7 @@ class TestMagic(unittest.TestCase):
             self.assertTrue(item.endswith(ext),
                             "Expected .{0}, got {1}".format(item.split(".")[1],
                                                             ext))
-                
+
     def test_media(self):
         """Test common media formats                    |"""
         for item in os.listdir(MEDIA_DIR):
@@ -134,7 +156,7 @@ class TestMagic(unittest.TestCase):
             self.assertTrue(item.endswith(ext),
                             "Expected .{0}, got {1}".format(item.split(".")[1],
                                                             ext))
-                
+
     def test_system(self):
         """Test common system formats                   |"""
         for item in os.listdir(SYSTEM_DIR):